TODO: Verify that tstools and audioselection was correctly merged

Conflicts:
	data/setup.xml
	lib/dvb/tstools.cpp
	lib/python/Components/Ipkg.py
	lib/python/Plugins/Extensions/GraphMultiEPG/plugin.py
	lib/python/Plugins/SystemPlugins/SoftwareManager/SoftwareTools.py
	lib/python/Plugins/SystemPlugins/SoftwareManager/plugin.py
	lib/python/Screens/AudioSelection.py
	lib/python/Screens/PluginBrowser.py
	main/bsod.cpp

+++ lib/service/servicemp3.h
@@ -164,7 +164,7 @@ public:
 		int bufferPercent;
 		int avgInRate;
 		int avgOutRate;
-		long long bufferingLeft;
+		int64_t bufferingLeft;
 		bufferInfo()
 			:bufferPercent(0), avgInRate(0), avgOutRate(0), bufferingLeft(-1)
 		{
